requirements.txt
Django>=6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=2.0
django-cors-headers>=4.0
